Fix for previous commit
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@166 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/Levels.h
@@ -22,7 +22,7 @@ class Player;
 
 namespace Levels {
 	extern int exps[200];
-	void giveEXP(Player* player, int exp, char type=0);
+	void giveEXP(Player* player, long exp, char type=0);
 	void addStat(Player* player, unsigned char* packet);
 	void setLevel(Player* player, int level);
 	void setJob(Player* player, int job);
